Put test AnnData helpers in the same file
