+++ tests/pna/pixeldataset/test_edgelist.py
@@ -7,34 +7,19 @@ import numpy as np
 import pandas as pd
 import polars as pl
 import pyarrow as pa
-from anndata import AnnData
 
 from pixelator.pna.pixeldataset import Edgelist, PNAPixelDataset
 from pixelator.pna.pixeldataset.types import Component
-
-
-class StubAnnDataHelper:
-    def __init__(self, adata: AnnData):
-        self._adata = adata
-        self.read_adata_calls: int = 0
-
-    def read_adata(
-        self, *, add_log1p_transform: bool, add_clr_transform: bool
-    ) -> AnnData:
-        self.read_adata_calls += 1
-        return self._adata
-
-
-def _make_adata(components: list[str], markers: list[str], x: np.ndarray) -> AnnData:
-    obs = pd.DataFrame(index=pd.Index(components, name="component"))
-    var = pd.DataFrame(index=pd.Index(markers, name="marker_id"))
-    return AnnData(X=x, obs=obs, var=var)
+from tests.pna.pixeldataset.ann_data_test_helpers import (
+    StubAnnDataHelper,
+    make_test_adata,
+)
 
 
 class TestEdgelistHelperInjection:
     def test_components_derived_from_injected_helper(self):
         components = ["c1", "c2"]
-        adata = _make_adata(components, ["m1", "m2"], x=np.array([[1, 2], [3, 4]]))
+        adata = make_test_adata(components, ["m1", "m2"], x=np.array([[1, 2], [3, 4]]))
         helper = StubAnnDataHelper(adata)
 
         edgelist = Edgelist(view=None, components=None, adata_helper=helper)
@@ -43,7 +28,9 @@ class TestEdgelistHelperInjection:
         assert helper.read_adata_calls >= 1
 
     def test_explicit_components_bypass_helper(self):
-        adata = _make_adata(["c1", "c2"], ["m1", "m2"], x=np.array([[1, 2], [3, 4]]))
+        adata = make_test_adata(
+            ["c1", "c2"], ["m1", "m2"], x=np.array([[1, 2], [3, 4]])
+        )
         helper = StubAnnDataHelper(adata)
 
         edgelist = Edgelist(view=None, components={"c1"}, adata_helper=helper)
