+++ tests/pna/pixeldataset/test_precomputed_layouts.py
@@ -6,33 +6,18 @@ Copyright © 2026 Pixelgen Technologies AB.
 import numpy as np
 import pandas as pd
 import polars as pl
-from anndata import AnnData
 
 from pixelator.pna.pixeldataset import PNAPixelDataset, PreComputedLayouts
-
-
-class StubAnnDataHelper:
-    def __init__(self, adata: AnnData):
-        self._adata = adata
-        self.read_adata_calls: int = 0
-
-    def read_adata(
-        self, *, add_log1p_transform: bool, add_clr_transform: bool
-    ) -> AnnData:
-        self.read_adata_calls += 1
-        return self._adata
-
-
-def _make_adata(components: list[str], markers: list[str], x: np.ndarray) -> AnnData:
-    obs = pd.DataFrame(index=pd.Index(components, name="component"))
-    var = pd.DataFrame(index=pd.Index(markers, name="marker_id"))
-    return AnnData(X=x, obs=obs, var=var)
+from tests.pna.pixeldataset.ann_data_test_helpers import (
+    StubAnnDataHelper,
+    make_test_adata,
+)
 
 
 class TestPreComputedLayoutsHelperInjection:
     def test_components_derived_from_injected_helper(self):
         components = ["c1", "c2"]
-        adata = _make_adata(components, ["m1", "m2"], x=np.array([[1, 2], [3, 4]]))
+        adata = make_test_adata(components, ["m1", "m2"], x=np.array([[1, 2], [3, 4]]))
         helper = StubAnnDataHelper(adata)
 
         layouts = PreComputedLayouts(
@@ -46,7 +31,9 @@ class TestPreComputedLayoutsHelperInjection:
         assert helper.read_adata_calls >= 1
 
     def test_explicit_components_bypass_helper(self):
-        adata = _make_adata(["c1", "c2"], ["m1", "m2"], x=np.array([[1, 2], [3, 4]]))
+        adata = make_test_adata(
+            ["c1", "c2"], ["m1", "m2"], x=np.array([[1, 2], [3, 4]])
+        )
         helper = StubAnnDataHelper(adata)
 
         layouts = PreComputedLayouts(
